fix: accept float for execution_order in reorder endpoint
Root cause of reordering issue:
- Frontend sends decimal values like 0.5, 1.5 for reordering
- Backend parameter was typed as Optional[int]
- FastAPI silently rejected/ignored float values
- Database never got updated with new execution_order
- Scripts stayed in same order

Fix:
- Change execution_order parameter from Optional[int] to Optional[float]
- Now accepts decimal values for intermediate positioning
- Reordering logic uses fractional orders (order - 1.5, order + 1.5)
- This allows inserting between existing integer orders without recalculating all

Example:
- Scripts: [order 1, order 2, order 3]
- Move order 2 up → newOrder = 2 - 1.5 = 0.5
- Result: [order 0.5, order 1, order 3]
- Next fetch sorts by order: 0.5 comes first ✓

+++ app/api/scripts_library.py
@@ -549,7 +549,7 @@ async def assign_script_to_repository(
 async def update_repository_script_assignment(
     repository_id: int,
     repo_script_id: int,
-    execution_order: Optional[int] = None,
+    execution_order: Optional[float] = None,
     enabled: Optional[bool] = None,
     custom_timeout: Optional[int] = None,
     custom_run_on: Optional[str] = None,
